#624 extract OyFallbackSwap
